More cleaning to get rid of linearalgebra code.
Part of #219
