+++ src/act/forcefield/forcefield_tables.cpp
@@ -39,7 +39,6 @@
 #include <cstdlib>
 
 #include "gromacs/fileio/xvgr.h"
-#include "gromacs/linearalgebra/matrix.h"
 #include "gromacs/math/utilities.h"
 #include "gromacs/math/vec.h"
 #include "gromacs/statistics/statistics.h"
