+++ src/gromacs/listed-forces/orires.cpp
@@ -42,9 +42,9 @@
 #include <cmath>
 
 #include "gromacs/gmxlib/network.h"
-#include "gromacs/linearalgebra/nrjac.h"
 #include "gromacs/math/do_fit.h"
 #include "gromacs/math/functions.h"
+#include "gromacs/math/nrjac.h"
 #include "gromacs/math/vec.h"
 #include "gromacs/mdtypes/commrec.h"
 #include "gromacs/mdtypes/fcdata.h"
