+++ src/gromacs/math/do_fit.cpp
@@ -41,8 +41,8 @@
 #include <cmath>
 #include <cstdio>
 
-#include "gromacs/linearalgebra/nrjac.h"
 #include "gromacs/math/functions.h"
+#include "gromacs/math/nrjac.h"
 #include "gromacs/math/utilities.h"
 #include "gromacs/math/vec.h"
 #include "gromacs/utility/fatalerror.h"
