+++ src/gromacs/math/nrjac.cpp
